Change Heaxagonalgrid to receive dimensions through constructor -including height scaling
HexagonalGrid now requests a Vector2 in its constructor. For hexagons
its only neccessary to now either the height or width as the other one
can be computed. However, instead of figuring out which is passed,
HexagonalGrid now just requests both

HexagonalGrid now requests a Vector2 in its constructor. For hexagons
its only neccessary to now either the height or width as the other one
can be computed. However, instead of figuring out which is passed,
HexagonalGrid now just requests both.