Pass instantiated objects to map
Previously we passed in only the prefabs to the HexGrid class. This
worked fine, but came with the problem that every component including
scripts and materials are changed for that very prefab. This makes
storing Neighbours per TileComponent impossible as they are stored for
the prefab but not the concrete instance of the tile. Now, we are
passing copies of the prefabs, so that each tile is unique. This comes
with the problem that they are visible before calling map.Show(), but
all at (0,0,0).