Find Storage seperately from Worker
This is a bugfix. Due to Unity immediatly enabling all scripts on a
GameObject, there is no way to set up a new worker with a Home prior to
the Resource Requesting Happiness Criteria. This criteria however
requires the worker to have a home in order to find the storage from
which to take the resources. WORKAROUND for now is that the requirement
finds the default storage on the GameManager empty.