Replace doc string comment with xml comment
XML comments show up in the preview and documentation features of most
IDEs outside of the code. Most standard code comments e.g.  '//', are
ignored outside of the actual source code. In Visual Studio, there is
the option to generate XML comments from template by triple-slashing '///' in
front of a method.